cluster: refactor cross cluster pre-deploy checks
