+++ pkg/cluster/spec/spec.go
@@ -25,7 +25,6 @@ import (
 	"github.com/pingcap/tiup/pkg/cluster/clusterutil"
 	"github.com/pingcap/tiup/pkg/cluster/template/scripts"
 	"github.com/pingcap/tiup/pkg/meta"
-	"github.com/pingcap/tiup/pkg/set"
 	"go.etcd.io/etcd/clientv3"
 )
 
@@ -36,12 +35,9 @@ const (
 
 // general role names
 var (
-	RoleMonitor              = "monitor"
-	RoleTiSparkMaster        = "tispark-master"
-	RoleTiSparkWorker        = "tispark-worker"
-	ErrNoTiSparkMaster       = errors.New("there must be a Spark master node if you want to use the TiSpark component")
-	ErrMultipleTiSparkMaster = errors.New("a TiSpark enabled cluster with more than 1 Spark master node is not supported")
-	ErrMultipleTisparkWorker = errors.New("multiple TiSpark workers on the same host is not supported by Spark")
+	RoleMonitor       = "monitor"
+	RoleTiSparkMaster = "tispark-master"
+	RoleTiSparkWorker = "tispark-worker"
 )
 
 type (
@@ -254,376 +250,6 @@ func findField(v reflect.Value, fieldName string) (int, bool) {
 		}
 	}
 	return -1, false
-}
-
-// platformConflictsDetect checks for conflicts in topology for different OS / Arch
-// set to the same host / IP
-func (s *Specification) platformConflictsDetect() error {
-	type (
-		conflict struct {
-			os   string
-			arch string
-			cfg  string
-		}
-	)
-
-	platformStats := map[string]conflict{}
-	topoSpec := reflect.ValueOf(s).Elem()
-	topoType := reflect.TypeOf(s).Elem()
-
-	for i := 0; i < topoSpec.NumField(); i++ {
-		if isSkipField(topoSpec.Field(i)) {
-			continue
-		}
-
-		compSpecs := topoSpec.Field(i)
-		for index := 0; index < compSpecs.Len(); index++ {
-			compSpec := compSpecs.Index(index)
-			// skip nodes imported from TiDB-Ansible
-			if compSpec.Interface().(InstanceSpec).IsImported() {
-				continue
-			}
-			// check hostname
-			host := compSpec.FieldByName("Host").String()
-			cfg := strings.Split(topoType.Field(i).Tag.Get("yaml"), ",")[0] // without meta
-			if host == "" {
-				return errors.Errorf("`%s` contains empty host field", cfg)
-			}
-
-			// platform conflicts
-			stat := conflict{
-				cfg: cfg,
-			}
-			if j, found := findField(compSpec, "OS"); found {
-				stat.os = compSpec.Field(j).String()
-			}
-			if j, found := findField(compSpec, "Arch"); found {
-				stat.arch = compSpec.Field(j).String()
-			}
-
-			prev, exist := platformStats[host]
-			if exist {
-				if prev.os != stat.os || prev.arch != stat.arch {
-					return &meta.ValidateErr{
-						Type:   meta.TypeMismatch,
-						Target: "platform",
-						LHS:    fmt.Sprintf("%s:%s/%s", prev.cfg, prev.os, prev.arch),
-						RHS:    fmt.Sprintf("%s:%s/%s", stat.cfg, stat.os, stat.arch),
-						Value:  host,
-					}
-				}
-			}
-			platformStats[host] = stat
-		}
-	}
-	return nil
-}
-
-func (s *Specification) portConflictsDetect() error {
-	type (
-		usedPort struct {
-			host string
-			port int
-		}
-		conflict struct {
-			tp  string
-			cfg string
-		}
-	)
-
-	portTypes := []string{
-		"Port",
-		"StatusPort",
-		"PeerPort",
-		"ClientPort",
-		"WebPort",
-		"TCPPort",
-		"HTTPPort",
-		"ClusterPort",
-	}
-
-	portStats := map[usedPort]conflict{}
-	uniqueHosts := set.NewStringSet()
-	topoSpec := reflect.ValueOf(s).Elem()
-	topoType := reflect.TypeOf(s).Elem()
-
-	for i := 0; i < topoSpec.NumField(); i++ {
-		if isSkipField(topoSpec.Field(i)) {
-			continue
-		}
-
-		compSpecs := topoSpec.Field(i)
-		for index := 0; index < compSpecs.Len(); index++ {
-			compSpec := compSpecs.Index(index)
-			// skip nodes imported from TiDB-Ansible
-			if compSpec.Interface().(InstanceSpec).IsImported() {
-				continue
-			}
-			// check hostname
-			host := compSpec.FieldByName("Host").String()
-			cfg := strings.Split(topoType.Field(i).Tag.Get("yaml"), ",")[0] // without meta
-			if host == "" {
-				return errors.Errorf("`%s` contains empty host field", cfg)
-			}
-			uniqueHosts.Insert(host)
-
-			// Ports conflicts
-			for _, portType := range portTypes {
-				if j, found := findField(compSpec, portType); found {
-					item := usedPort{
-						host: host,
-						port: int(compSpec.Field(j).Int()),
-					}
-					tp := compSpec.Type().Field(j).Tag.Get("yaml")
-					prev, exist := portStats[item]
-					if exist {
-						return &meta.ValidateErr{
-							Type:   meta.TypeConflict,
-							Target: "port",
-							LHS:    fmt.Sprintf("%s:%s.%s", prev.cfg, item.host, prev.tp),
-							RHS:    fmt.Sprintf("%s:%s.%s", cfg, item.host, tp),
-							Value:  item.port,
-						}
-					}
-					portStats[item] = conflict{
-						tp:  tp,
-						cfg: cfg,
-					}
-				}
-			}
-		}
-	}
-
-	// Port conflicts in monitored components
-	monitoredPortTypes := []string{
-		"NodeExporterPort",
-		"BlackboxExporterPort",
-	}
-	monitoredOpt := topoSpec.FieldByName(monitorOptionTypeName)
-	for host := range uniqueHosts {
-		cfg := "monitored"
-		for _, portType := range monitoredPortTypes {
-			f := monitoredOpt.FieldByName(portType)
-			item := usedPort{
-				host: host,
-				port: int(f.Int()),
-			}
-			ft, found := monitoredOpt.Type().FieldByName(portType)
-			if !found {
-				return errors.Errorf("incompatible change `%s.%s`", monitorOptionTypeName, portType)
-			}
-			// `yaml:"node_exporter_port,omitempty"`
-			tp := strings.Split(ft.Tag.Get("yaml"), ",")[0]
-			prev, exist := portStats[item]
-			if exist {
-				return &meta.ValidateErr{
-					Type:   meta.TypeConflict,
-					Target: "port",
-					LHS:    fmt.Sprintf("%s:%s.%s", prev.cfg, item.host, prev.tp),
-					RHS:    fmt.Sprintf("%s:%s.%s", cfg, item.host, tp),
-					Value:  item.port,
-				}
-			}
-			portStats[item] = conflict{
-				tp:  tp,
-				cfg: cfg,
-			}
-		}
-	}
-
-	return nil
-}
-
-func (s *Specification) dirConflictsDetect() error {
-	type (
-		usedDir struct {
-			host string
-			dir  string
-		}
-		conflict struct {
-			tp       string
-			cfg      string
-			imported bool
-		}
-	)
-
-	dirTypes := []string{
-		"DataDir",
-		"DeployDir",
-	}
-
-	// usedInfo => type
-	var dirStats = map[usedDir]conflict{}
-
-	topoSpec := reflect.ValueOf(s).Elem()
-	topoType := reflect.TypeOf(s).Elem()
-
-	for i := 0; i < topoSpec.NumField(); i++ {
-		if isSkipField(topoSpec.Field(i)) {
-			continue
-		}
-
-		compSpecs := topoSpec.Field(i)
-		for index := 0; index < compSpecs.Len(); index++ {
-			compSpec := compSpecs.Index(index)
-			// check hostname
-			host := compSpec.FieldByName("Host").String()
-			cfg := strings.Split(topoType.Field(i).Tag.Get("yaml"), ",")[0] // without meta
-			if host == "" {
-				return errors.Errorf("`%s` contains empty host field", cfg)
-			}
-
-			// Directory conflicts
-			for _, dirType := range dirTypes {
-				if j, found := findField(compSpec, dirType); found {
-					item := usedDir{
-						host: host,
-						dir:  compSpec.Field(j).String(),
-					}
-					// data_dir is relative to deploy_dir by default, so they can be with
-					// same (sub) paths as long as the deploy_dirs are different
-					if item.dir != "" && !strings.HasPrefix(item.dir, "/") {
-						continue
-					}
-					// `yaml:"data_dir,omitempty"`
-					tp := strings.Split(compSpec.Type().Field(j).Tag.Get("yaml"), ",")[0]
-					prev, exist := dirStats[item]
-					// not checking between imported nodes
-					if exist &&
-						!(compSpec.Interface().(InstanceSpec).IsImported() && prev.imported) {
-						return &meta.ValidateErr{
-							Type:   meta.TypeConflict,
-							Target: "directory",
-							LHS:    fmt.Sprintf("%s:%s.%s", prev.cfg, item.host, prev.tp),
-							RHS:    fmt.Sprintf("%s:%s.%s", cfg, item.host, tp),
-							Value:  item.dir,
-						}
-					}
-					// not reporting error for nodes imported from TiDB-Ansible, but keep
-					// their dirs in the map to check if other nodes are using them
-					dirStats[item] = conflict{
-						tp:       tp,
-						cfg:      cfg,
-						imported: compSpec.Interface().(InstanceSpec).IsImported(),
-					}
-				}
-			}
-		}
-	}
-
-	return nil
-}
-
-// CountDir counts for dir paths used by any instance in the cluster with the same
-// prefix, useful to find potential path conflicts
-func (s *Specification) CountDir(targetHost, dirPrefix string) int {
-	dirTypes := []string{
-		"DataDir",
-		"DeployDir",
-		"LogDir",
-	}
-
-	// path -> count
-	dirStats := make(map[string]int)
-	count := 0
-	topoSpec := reflect.ValueOf(s).Elem()
-	dirPrefix = clusterutil.Abs(s.GlobalOptions.User, dirPrefix)
-
-	for i := 0; i < topoSpec.NumField(); i++ {
-		if isSkipField(topoSpec.Field(i)) {
-			continue
-		}
-
-		compSpecs := topoSpec.Field(i)
-		for index := 0; index < compSpecs.Len(); index++ {
-			compSpec := compSpecs.Index(index)
-			// Directory conflicts
-			for _, dirType := range dirTypes {
-				if j, found := findField(compSpec, dirType); found {
-					dir := compSpec.Field(j).String()
-					host := compSpec.FieldByName("Host").String()
-
-					switch dirType { // the same as in instance.go for (*instance)
-					case "DataDir":
-						deployDir := compSpec.FieldByName("DeployDir").String()
-						// the default data_dir is relative to deploy_dir
-						if dir != "" && !strings.HasPrefix(dir, "/") {
-							dir = filepath.Join(deployDir, dir)
-						}
-					case "LogDir":
-						deployDir := compSpec.FieldByName("DeployDir").String()
-						field := compSpec.FieldByName("LogDir")
-						if field.IsValid() {
-							dir = field.Interface().(string)
-						}
-
-						if dir == "" {
-							dir = "log"
-						}
-						if !strings.HasPrefix(dir, "/") {
-							dir = filepath.Join(deployDir, dir)
-						}
-					}
-					dir = clusterutil.Abs(s.GlobalOptions.User, dir)
-					dirStats[host+dir]++
-				}
-			}
-		}
-	}
-
-	for k, v := range dirStats {
-		if k == targetHost+dirPrefix || strings.HasPrefix(k, targetHost+dirPrefix+"/") {
-			count += v
-		}
-	}
-
-	return count
-}
-
-func (s *Specification) validateTiSparkSpec() error {
-	// There must be a Spark master
-	if len(s.TiSparkMasters) == 0 {
-		if len(s.TiSparkWorkers) == 0 {
-			return nil
-		}
-		return ErrNoTiSparkMaster
-	}
-
-	// We only support 1 Spark master at present
-	if len(s.TiSparkMasters) > 1 {
-		return ErrMultipleTiSparkMaster
-	}
-
-	// Multiple workers on the same host is not supported by Spark
-	if len(s.TiSparkWorkers) > 1 {
-		cnt := make(map[string]int)
-		for _, w := range s.TiSparkWorkers {
-			if cnt[w.Host] > 0 {
-				return errors.Annotatef(ErrMultipleTisparkWorker, "the host %s is duplicated", w.Host)
-			}
-			cnt[w.Host]++
-		}
-	}
-
-	return nil
-}
-
-// Validate validates the topology specification and produce error if the
-// specification invalid (e.g: port conflicts or directory conflicts)
-func (s *Specification) Validate() error {
-	if err := s.platformConflictsDetect(); err != nil {
-		return err
-	}
-
-	if err := s.portConflictsDetect(); err != nil {
-		return err
-	}
-
-	if err := s.dirConflictsDetect(); err != nil {
-		return err
-	}
-
-	return s.validateTiSparkSpec()
 }
 
 // GetPDList returns a list of PD API hosts of the current cluster
