+++ pkg/cluster/spec/spec_manager.go
@@ -173,6 +173,23 @@ func (s *SpecManager) List() (names []string, err error) {
 	return
 }
 
+// GetAllClusters get a metadata list of all clusters deployed by current user
+func (s *SpecManager) GetAllClusters() (clusters map[string]Metadata, err error) {
+	names, err := s.List()
+	if err != nil {
+		return nil, errors.AddStack(err)
+	}
+	for _, name := range names {
+		metadata := s.NewMetadata()
+		err = s.Metadata(name, metadata)
+		if err != nil {
+			return nil, errors.Trace(err)
+		}
+		clusters[name] = metadata
+	}
+	return
+}
+
 // ensureDir ensures that the cluster directory exists.
 func (s *SpecManager) ensureDir(clusterName string) error {
 	if err := utils.CreateDir(s.Path(clusterName)); err != nil {
